scripts/requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
google-generativeai>=0.3.0
torch
transformers
umap-learn
scikit-learn
tqdm
numpy
